+++ libs/config/hapify/src/validators/validators.interface.d.ts
@@ -1,0 +1,1 @@
+export type ValidatorReturn = { errors?: string[]; warnings?: string[] };