+++ libs/nestjs/core/src/decorators/abilities.decorator.ts
@@ -1,22 +1,11 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { GqlExecutionContext } from '@nestjs/graphql';
+
+import { getRequestFromContext } from '../helpers';
 
 /**
  * Get the abilities from the request
  */
 export const CurrentAbilities = createParamDecorator(
-  (_data: unknown, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    return request.abilities;
-  },
-);
-
-/**
- * Get the abilities from the request in GraphQL
- */
-export const CurrentGqlAbilities = createParamDecorator(
-  (_data: unknown, context: ExecutionContext) => {
-    const ctx = GqlExecutionContext.create(context);
-    return ctx.getContext().req.abilities;
-  },
+  (_data: unknown, context: ExecutionContext) =>
+    getRequestFromContext(context).abilities,
 );